Add caching for PSP data
Redis is queried to see if the PSP data (stored as JSON) exists for each of
the search result ids. For the ones that don't exist, the db will be queried
only for those specific PSPs. They are then merged with the cached PSPs
while keeping the search result order.