Add function to determine if the stock market is open or not
The results are obtained from the Alpaca API and cached into Redis.
The cache data expires at market open if the market is currently closed.
Likewise, if the market is currently opened, the data expires at market close.
Market closures due to the weekends are accounted for. Market hours cannot
be assumed to always be Mon-Fri 9:30am-4:00pm Eastern due to closing or
shortening market hours for holidays.