Add caching to stock prices
If the stock market is open, we don't want to cache prices since they can
change rapidly and the cached data will become stale. We only cache if the
market is closed, since prices cannot change. For any cache misses, we will
query the Alpaca API then cache this data.